close gh-4, gh-5, gh-12, gh-11, gh-9, gh-10

+++ src/main/java/com/rivertech/betgametask/bet/Bet.java
@@ -44,7 +44,7 @@ public class Bet implements Serializable {
 
     private Long betAmount;
 
-    private double wonAmount;
+    private Long wonAmount;
 
     private int betNum;
